@@ -10,16 +10,16 @@
   "license": "ISC",
   "devDependencies": {
     "babel-core": "^6.0.20",
-    "babel-eslint": "^4.1.3",
+    "babel-eslint": "^6.0.4",
     "babel-loader": "^6.0.1",
     "babel-preset-es2015": "^6.0.15",
     "babel-preset-react": "^6.0.15",
     "babel-preset-stage-0": "^6.0.15",
     "css-loader": "^0.23.1",
-    "eslint-plugin-react": "^3.6.2",
-    "install": "^0.4.1",
+    "eslint-plugin-react": "^5.1.1",
+    "install": "^0.8.1",
     "npm": "^3.5.3",
-    "postcss-loader": "^0.8.0",
+    "postcss-loader": "^0.9.1",
     "precss": "^1.4.0",
     "react-hot-loader": "^1.3.0",
     "style-loader": "^0.13.0",
@@ -28,7 +28,9 @@
   },
   "dependencies": {
     "classnames": "^2.2.3",
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0"
+    "flux": "^2.0.1",
+    "react": "^15.1.0",
+    "react-addons-pure-render-mixin": "^15.1.0",
+    "react-dom": "^15.1.0"
   }
 }
