@@ -5,8 +5,7 @@ import classNames from 'classnames';
 export default class TextFilter extends Component {
 
   static propTypes:{
-    column: React.PropTypes.object,
-    filterChangeCallback: React.PropTypes.func
+    column: React.PropTypes.object
     };
 
   constructor(props) {
@@ -19,11 +18,10 @@ export default class TextFilter extends Component {
 
 
   onFilterChange = (e)=> {
-    e.preventDefault();
     var value = e.target.value;
     this.setState({filterValue: value});
     var filter = this.buildFilter(value);
-    this.props.filterChangeCallback(filter);
+    //this.props.filterChangeCallback(filter);
   };
 
 
